Theoretically, at least, loading a process but not executing it yet
