+++ andy_elf/mmgr/process.h
@@ -17,21 +17,24 @@
 #define PROCESS_TABLE_ENTRY_SIG 0x54504552
 
 struct ProcessTableEntry {
-  uint32_t magic; //must be 0x54504552 = 'PTBE'
+  uint32_t magic; //must be 0x54504552 = 'PTBE'               //offset 0x0
   //saved register states are on the process's stack
-  size_t esp;
-  uint16_t ecs;
-  size_t stack_page_count;
+  size_t esp;                                                 //offset 0x04
+  uint16_t ecs;                                               //offset 0x08
+  size_t stack_page_count;                                    //offset 0x0A 
 
   //process state
-  uint8_t status;
+  uint8_t status;                                             //offset 0x0E
+  uint8_t unused;                                             //offset 0x0F
 
   //memory location and stats
-  void* root_paging_directory_phys; //physical RAM address of the root paging directory
-  void* root_paging_directory_kmem; //virtual RAM address of the root paging directory, in kernel space.
-  void* heap_start;                 //location of the heap, in process space
-  size_t heap_allocated;
-  size_t heap_used;
+  void* root_paging_directory_phys; //physical RAM address of the root paging directory. Offset 0x10
+  void* root_paging_directory_kmem; //virtual RAM address of the root paging directory, in kernel space. NULL if not currently mapped. Offset 0x14.
+  void* heap_start;                 //location of the heap, in process space. Offset 0x18
+  size_t heap_allocated;            //offset 0x1D
+  size_t heap_used;                 //offset 0x22
+  void *stack_phys_ptr;             //offset 0x26
+  uint32_t *stack_kmem_ptr;         //offset 0x2A
 } __attribute__((packed));
 
 void initialise_process_table(uint32_t* kernel_paging_directory);
