+++ andy_elf/scheduler/Makefile
@@ -4,7 +4,10 @@ all: libscheduler.a
 
 clean:
 	rm -f *.o *.a
-	
+
+lowlevel.o:
+	nasm -f elf32 lowlevel.asm
+
 process.o: process.c process.h
 	gcc -c ${CFLAGS} process.c
 
@@ -14,5 +17,5 @@ scheduler_task.o: scheduler_task.c scheduler_task_internals.h ../include/schedul
 scheduler.o: scheduler.c scheduler_task_internals.h ../include/scheduler/scheduler.h
 		gcc -c ${CFLAGS} scheduler.c
 
-libscheduler.a: scheduler.o scheduler_task.o
-	ar rcs libscheduler.a scheduler.o scheduler_task.o
+libscheduler.a: scheduler.o scheduler_task.o lowlevel.o
+	ar rcs libscheduler.a scheduler.o scheduler_task.o lowlevel.o
